♻️ Refactor CLI command structure for fgt get cmd firewall commands
