+++ fotoobo/cli/fgt/get/cmdb/firewall/firewall.py
@@ -9,16 +9,10 @@ import typer
 
 from fotoobo.exceptions import GeneralError
 from fotoobo.helpers import cli_path
-from fotoobo.tools.fgt.cmdb.firewall.address import get_firewall_address
-from fotoobo.tools.fgt.cmdb.firewall.addrgrp import get_firewall_addrgrp
-
-from .service import service
+from fotoobo.tools.fgt.cmdb.firewall import *  # pylint: disable=wildcard-import
 
 app = typer.Typer(no_args_is_help=True, rich_markup_mode="rich")
 log = logging.getLogger("fotoobo")
-
-
-app.add_typer(service.app, name="service", help="FortiGate get cmdb firewall service commands.")
 
 
 @app.callback()
@@ -68,7 +62,7 @@ def address(
     if name and ("*" in vdom or "," in vdom):
         raise GeneralError("With name argument you have to specify one single VDOM (with --vdom)")
 
-    result = get_firewall_address(host, name, vdom, output_file)
+    result = get_cmdb_firewall_address(host, name, vdom, output_file)
     if not output_file:
         result.print_table_raw(result.results[host], auto_header=True, title=host)
 
@@ -108,6 +102,86 @@ def addrgrp(
     if name and ("*" in vdom or "," in vdom):
         raise GeneralError("With name argument you have to specify one single VDOM (with --vdom)")
 
-    result = get_firewall_addrgrp(host, name, vdom, output_file)
+    result = get_cmdb_firewall_addrgrp(host, name, vdom, output_file)
+    if not output_file:
+        result.print_table_raw(result.results[host], auto_header=True, title=host)
+
+
+@app.command()
+def service_custom(
+    host: str = typer.Argument(
+        "",
+        help="The FortiGate hostname to access (must be defined in the inventory). "
+        "\[default: <all>]",
+        show_default=False,
+        metavar="[host]",
+    ),
+    name: str = typer.Argument(
+        "",
+        help="The firewall service object to get \[default: <all>]",
+        show_default=False,
+        metavar="[name]",
+    ),
+    vdom: str = typer.Option(
+        "*",
+        "--vdom",
+        help="The vdom to query ('vdom1' or 'vdom1,vdom2') \[default: <all>]",
+        show_default=False,
+        metavar="[vdom]",
+    ),
+    output_file: str = typer.Option(
+        None,
+        "--output",
+        "-o",
+        help="Output file (format is specified by extension)",
+        metavar="[file]",
+        show_default=False,
+    ),
+) -> None:
+    """Get FortiGate cmdb firewall service custom."""
+    if name and ("*" in vdom or "," in vdom):
+        raise GeneralError("With name argument you have to specify one single VDOM (with --vdom)")
+
+    result = get_cmdb_firewall_service_custom(host, name, vdom, output_file)
+    if not output_file:
+        result.print_table_raw(result.results[host], auto_header=True, title=host)
+
+
+@app.command()
+def service_group(
+    host: str = typer.Argument(
+        "",
+        help="The FortiGate hostname to access (must be defined in the inventory). "
+        "\[default: <all>]",
+        show_default=False,
+        metavar="[host]",
+    ),
+    name: str = typer.Argument(
+        "",
+        help="The firewall service group to get \[default: <all>]",
+        show_default=False,
+        metavar="[name]",
+    ),
+    vdom: str = typer.Option(
+        "*",
+        "--vdom",
+        help="The vdom to query ('vdom1' or 'vdom1,vdom2') \[default: <all>]",
+        show_default=False,
+        metavar="[vdom]",
+    ),
+    output_file: str = typer.Option(
+        None,
+        "--output",
+        "-o",
+        help="Output file (format is specified by extension)",
+        metavar="[file]",
+        show_default=False,
+    ),
+) -> None:
+    """Get FortiGate cmdb firewall service group."""
+    if name and ("*" in vdom or "," in vdom):
+        raise GeneralError("With name argument you have to specify one single VDOM (with --vdom)")
+
+    result = get_cmdb_firewall_service_group(host, name, vdom, output_file)
     if not output_file:
         result.print_table_raw(result.results[host], auto_header=True, title=host)
