--- fotoobo/cli/fgt/get/cmdb/firewall/service/__init__.py
@@ -1,7 +1,0 @@
-"""
-__init__.py
-"""
-
-from . import service
-
-__all__ = ["service"]
