+++ fotoobo/tools/fgt/cmdb/firewall/address.py
@@ -7,7 +7,9 @@ from fotoobo.helpers.result import Result
 from fotoobo.tools.fgt.get import api
 
 
-def get_firewall_address(host: str, name: str, vdom: str, output_file: str) -> Result[list[Any]]:
+def get_cmdb_firewall_address(
+    host: str, name: str, vdom: str, output_file: str
+) -> Result[list[Any]]:
     """Get the firewall address object(s)
 
     The FortiGate api endpoint is: /cmdb/firewall/address
